requirements.txt
Flask==2.0.1
Flask-Migrate==4.0.0
Flask-SQLAlchemy==2.5.1
joblib==1.4.2
numpy==1.22.3
xgboost==2.1.0
gunicorn==20.1.0
Werkzeug==2.0.3
scikit-learn==1.2.2
pandas==1.4.3
boto3==1.24.0
psycopg2-binary==2.9.7
SQLAlchemy==1.4.46
